rename acrawriter -> acra-writer / AcraWriter
