+++ benchmarks/write/write.go
@@ -19,7 +19,7 @@ package write
 import (
 	"database/sql"
 	"fmt"
-	"github.com/cossacklabs/acra/acrawriter"
+	"github.com/cossacklabs/acra/acra-writer"
 	"github.com/cossacklabs/acra/benchmarks/common"
 	"github.com/cossacklabs/acra/benchmarks/config"
 	"github.com/cossacklabs/acra/utils"
@@ -49,7 +49,7 @@ func CheckOneKey() {
 	}
 }
 
-// GetPublicOneKey load and return public key for acrawriter <onekey_storage.pub>
+// GetPublicOneKey load and return public key for acra-writer <onekey_storage.pub>
 func GetPublicOneKey() *keys.PublicKey {
 	publicKey, err := utils.LoadPublicKey("src/github.com/cossacklabs/acra/benchmarks/.acrakeys/onekey_storage.pub")
 	if err != nil {
