+++ examples/golang/src/example_with_zone/example_with_zone.go
@@ -19,7 +19,7 @@ import (
 	"database/sql"
 	"encoding/json"
 	"fmt"
-	"github.com/cossacklabs/acra/acrawriter"
+	"github.com/cossacklabs/acra/acra-writer"
 	"github.com/cossacklabs/themis/gothemis/keys"
 	_ "github.com/lib/pq"
 	log "github.com/sirupsen/logrus"
