+++ wrappers/ruby/acrawriter/test/test_acrawriter.rb
@@ -2,7 +2,7 @@ require 'minitest/autorun'
 require 'rubythemis'
 require 'acrawriter'
 
-class TestAcrawriter < Minitest::Test
+class TestAcraWriter < Minitest::Test
   def test_create_acrastruct
     generator = Themis::SKeyPairGen.new
     private, public = generator.ec
